requirements.txt
absl-py==0.7.1
astor==0.7.1
astroid==2.1.0
autopep8==1.4.3
backcall==0.1.0
beautifulsoup4==4.6.3
bleach==2.1.4
bs4==0.0.1
cachetools==3.1.1
certifi==2018.8.13
chardet==3.0.4
Click==7.0
colorama==0.3.9
cycler==0.10.0
decorator==4.3.0
dialogflow==0.6.0
Django==2.1.1
entrypoints==0.2.3
et-xmlfile==1.0.1
Flask==1.0.2
gast==0.2.2
google==2.0.2
google-api-core==1.13.0
google-auth==1.6.3
googleapis-common-protos==1.6.0
grpcio==1.19.0
gTTS==2.0.3
gTTS-token==1.1.3
h5py==2.9.0
html5lib==1.0.1
idna==2.7
imageio==2.5.0
ipykernel==4.8.2
ipython==6.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.0
isort==4.3.4
itsdangerous==1.1.0
jdcal==1.4
jedi==0.12.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
lxml==4.3.1
Markdown==3.1
MarkupSafe==1.0
matplotlib==2.2.3
mccabe==0.6.1
mistune==0.8.3
mock==2.0.0
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.3
notebook==5.6.0
numpy==1.16.2
omdb==0.10.0
opencv-python==4.0.0.21
openpyxl==2.6.1
paho-mqtt==1.4.0
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
pbr==5.1.3
pickleshare==0.7.4
Pillow==5.4.1
playsound==1.2.2
prometheus-client==0.3.1
prompt-toolkit==1.0.15
protobuf==3.7.0
psycopg2==2.7.7
pyasn1==0.4.5
pyasn1-modules==0.2.5
PyBluez==0.22
pycodestyle==2.5.0
pydot==1.4.1
pygame==1.9.6
Pygments==2.2.0
pylint==2.2.2
pyparsing==2.2.0
pypiwin32==223
PyQt5==5.11.2
PyQt5-sip==4.19.12
python-dateutil==2.7.3
python-vlc==3.0.6109
pytz==2018.5
PyWavelets==1.0.3
pywin32==224
pywinpty==0.5.4
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.1
requests==2.19.1
rsa==4.0
scikit-image==0.15.0
scikit-learn==0.19.2
scipy==1.1.0
seaborn==0.9.0
selenium==3.141.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
tornado==5.1
traitlets==4.3.2
urllib3==1.23
virtualenv==16.0.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.0
wrapt==1.11.1
xlrd==1.1.0
